work on patchWindow + project file managment

+++ src/frontend/SettingsWindow/LibraryManagerWindow/ListLightsWindow/ListLightsWindow.cpp
@@ -40,6 +40,16 @@ void ListLightsWindow::DrawContents()
         GLOBAL::ADDLIGHTWINDOW::cameFromListLightsWindow = true;
     }
 
+    if (GLOBAL::LISTLIGHTSWINDOW::activeItemIndex > 0 && (ImGui::IsKeyPressed(ImGuiKey_UpArrow) && GLOBAL::KEYHANDLER::isKeyDown_UpArrow)) {
+        GLOBAL::LISTLIGHTSWINDOW::activeItemIndex--;
+        GLOBAL::KEYHANDLER::isKeyDown_UpArrow = true;
+    }
+
+    if (GLOBAL::LISTLIGHTSWINDOW::activeItemIndex < GLOBAL::LIGHTFILEMANAGER::lightsLibrary.size() - 1 && (ImGui::IsKeyPressed(ImGuiKey_DownArrow) && GLOBAL::KEYHANDLER::isKeyDown_DownArrow)) {
+        GLOBAL::LISTLIGHTSWINDOW::activeItemIndex++;
+        GLOBAL::KEYHANDLER::isKeyDown_DownArrow = true;
+    }
+
 
     // if (GLOBAL::LIGHTFILEMANAGER::lightsLibrary.empty()) { ImGui::Text("No lights available."); }
 
@@ -120,8 +130,9 @@ void ListLightsWindow::DrawContents()
 
     ImGui::SetCursorPos(ImVec2(_size.x - saveMargin - ImGui::CalcTextSize("Cancel").x - 2 * saveMargin,
                                _size.y - ImGui::CalcTextSize("XXX").y - 3 * saveMargin));
-    if (ImGui::Button("Add", ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin))/* || ImGui::IsKeyPressed(ImGuiKey_A)*/)
+    if (ImGui::Button("Add", ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin)) || (ImGui::IsKeyPressed(ImGuiKey_A) && !GLOBAL::KEYHANDLER::isKeyDown_A))
     {
+        GLOBAL::KEYHANDLER::isKeyDown_A = true;
         GLOBAL::ADDLIGHTWINDOW::isWindowActive           = true;
         GLOBAL::LISTLIGHTSWINDOW::isWindowActive         = false;
         GLOBAL::ADDLIGHTWINDOW::cameFromListLightsWindow = true;
@@ -133,8 +144,9 @@ void ListLightsWindow::DrawContents()
 
     ImGui::SetCursorPos(ImVec2(_size.x - 2 * saveMargin - 2 * ImGui::CalcTextSize("Cancel").x - 4 * saveMargin,
                                _size.y - ImGui::CalcTextSize("XXX").y - 3 * saveMargin));
-    if (ImGui::Button("Edit", ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin))/* || ImGui::IsKeyPressed(ImGuiKey_E)*/)
+    if (ImGui::Button("Edit", ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin)) || (ImGui::IsKeyPressed(ImGuiKey_E) && !GLOBAL::KEYHANDLER::isKeyDown_E))
     {
+        GLOBAL::KEYHANDLER::isKeyDown_E = true;
         GLOBAL::ADDLIGHTWINDOW::lightToEdit = GLOBAL::LIGHTFILEMANAGER::lightsLibrary.at(GLOBAL::LISTLIGHTSWINDOW::activeItemIndex);
         GLOBAL::ADDLIGHTWINDOW::isWindowActive           = true;
         GLOBAL::LISTLIGHTSWINDOW::isWindowActive         = false;
@@ -149,8 +161,9 @@ void ListLightsWindow::DrawContents()
     ImGui::SetCursorPos(ImVec2(_size.x - 3 * saveMargin - 3 * ImGui::CalcTextSize("Cancel").x - 6 * saveMargin,
                                _size.y - ImGui::CalcTextSize("XXX").y - 3 * saveMargin));
     if (ImGui::Button("Delete",
-                      ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin))/* || ImGui::IsKeyPressed(ImGuiKey_Delete)*/)
+                      ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin)) || (ImGui::IsKeyPressed(ImGuiKey_Delete) && !GLOBAL::KEYHANDLER::isKeyDown_Delete))
     {
+        GLOBAL::KEYHANDLER::isKeyDown_Delete = true;
         // std::string lightName = GLOBAL::LIGHTFILEMANAGER::lightsLibrary.at(GLOBAL::LISTLIGHTSWINDOW::activeItemIndex).name;
         LightFileManager::deleteLightByIndex(GLOBAL::LISTLIGHTSWINDOW::activeItemIndex);
     }
@@ -167,8 +180,9 @@ void ListLightsWindow::DrawContents()
     // ImGui::CalcTextSize("XXX").y - 3 * saveMargin));
     ImGui::SetCursorPos(ImVec2(saveMargin, _size.y - ImGui::CalcTextSize("XXX").y - 3 * saveMargin));
     if (ImGui::Button("Cancel",
-                      ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin))/* || ImGui::IsKeyPressed(ImGuiKey_Escape)*/)
+                      ImVec2(ImGui::CalcTextSize("Cancel").x + 2 * saveMargin, ImGui::CalcTextSize("XXX").y + 2 * saveMargin)) || (ImGui::IsKeyPressed(ImGuiKey_Escape) && !GLOBAL::KEYHANDLER::isKeyDown_Escape))
     {
+        GLOBAL::KEYHANDLER::isKeyDown_Escape = true;
         GLOBAL::LISTLIGHTSWINDOW::isWindowActive         = false;
         GLOBAL::ADDLIGHTWINDOW::cameFromListLightsWindow = false;
     }
